+++ contentcuration/contentcuration/frontend/channelEdit/components/move/__tests__/newTopicModal.spec.js
@@ -1,6 +1,6 @@
 import { mount } from '@vue/test-utils';
 import NewTopicModal from '../NewTopicModal.vue';
-import store from '../../store';
+import store from '../../../store';
 
 function makeWrapper() {
   return mount(NewTopicModal, {
